+++ components/Util/ToggleProvider.tsx
@@ -1,6 +1,6 @@
 import { useToggle } from '@/hooks/useToggle';
 import { createUseContextHook } from '@/utils/createUseContextHook';
-import { createContext } from 'react';
+import { createContext, ReactElement } from 'react';
 import { PassProps } from './PassProps';
 import { useChildCount } from '@/hooks/useChildCount';
 import { useFirstChild } from '@/hooks/useFirstChild';
@@ -31,13 +31,14 @@ type TriggerProps = React.PropsWithChildren & { [x: string]: any };
 ToggleProvider.Trigger = function ({ children, ...props }: TriggerProps) {
   useChildCount(children, 1, 'ToggleProvider.Trigger expects exactly one child!');
   const { toggleState } = useToggleProviderContext();
-  const child = useFirstChild(children);
+  const child = useFirstChild(children) as ReactElement<any>;
 
   return (
     <PassProps
       {...props}
       onClick={(e: any) => {
         props.onClick && props.onClick(e);
+        child.props.onClick && child.props.onClick(e);
         toggleState();
       }}>
       {child}
@@ -48,13 +49,14 @@ ToggleProvider.Trigger = function ({ children, ...props }: TriggerProps) {
 type TargetProps = React.PropsWithChildren & { [x: string]: any };
 ToggleProvider.Target = function ({ children, ...props }: TargetProps) {
   useChildCount(children, 1, 'ToggleProvider.Target expects exactly one child!');
-  const { isToggled } = useToggleProviderContext();
+  const { isToggled, toggleState } = useToggleProviderContext();
   const child = useFirstChild(children);
 
   return (
     <PassProps
       {...props}
-      isToggled={isToggled}>
+      isToggled={isToggled}
+      onToggle={state => toggleState(state)}>
       {child}
     </PassProps>
   );